feat: implement turbo streams over signalr websockets
This is to imitate the `turbo-cable-stream-source`tag that rails
provides. It will now update any client connected with the updated list
of todos

This also moves the asset management over to libman so we can build a
self contained image and not depend on jsdelivr when we are running our
app.